Rename and move TextAreaPanelPlugin
